Develop (#23)
* Bot (#20)

* Add chatbot flows,
Add bot.qr.png to .gitignore

* Fetch bot menus and answers from DB

* Auth fix (#22)

* Fix double refresh hanging the server

* 🧹 Clean up

* Add console.error() to exitLog

* Testing (#24)

* Add role to player on PlayersDAO.upsert

* Add tests

* Alquimia (#25)

* Add alquimia payment verification

* Update seed.sql

* Comment out services.old.ts

* Remove Player from DepositDAO.create()

* Remove Player from DepositDAO.create()

* Alquimia (#26)

* Fix deposit test

* Look for deposits 24 hours into the past

* Fix tests

* Update deposit on /transactions/deposit/:id (#27)

* Alquimia sync (#28)

* Update deposit on /transactions/deposit/:id

* Sync deposits from Alquimia DB to local on FinanceServices.verifyPayment()

* Set deposit status to rejected if not found

* Change integer incremental IDs to string UUIDs

* Modify endpoints to use only GET and POST methods (#29)

* Modify endpoints to use only GET and POST methods

* Misc (#30)

* Add endpoint to show Alquimia bank details,
Configure cors for development,
Update deposit instead of create when existing tracking key found in request

* Add ALLOWED_ORIGIN to .env

* origin source

* lint

---------


---------


---------


* Deposit flow (#31)

* Create CasinoCoinsService to handle coin transfers,
Improve FinanceService.confirmDeposit error handling,
Remove paid_at and coins_transfered from Deposit model,
add casinoTokenService.login() call to seed.ts,
Improve Notify error handling,
Improve TransactionsDAO error handling,
Turn Player.balance_currency into a required field,
Swap Transaction.status with Transaction.ok

* Update .gitignore

* Fix bug in FinanceServices.alquimiaDepositLookup (#32)

* Fix bug in FinanceServices.alquimiaDepositLookup

* Improve bank-account request validator

* Fix bank-account test

* fix cashout validation

* Add rate limiter to POST /transactions/deposit/:id? (#33)

* Deposit (#34)

* Improve DepositDAO.authorizeConfirmation,
Give agent role permission to confirm deposit,
Remove TransactionsController.deleteDeposit,
Fixed finalizeDeposit logic to not send coins on confirmed deposits,
Add optional id to GET /agent/depoists to retreive individual depoist,
Rename /agent/deposits/complete to /agent/pending/depoists

* Update README

* Payment throttling (#35)

* Update dependecies,
Restrict withdrawals to 1 every 24 hours

* Update CHANGELOG

* update bankNumber validation

* Logging (#36)

* Add logtailLogger,
Fix bug in AuthServices.refresh

* Update CHANGELOG

* Update CONFIG.LOG.CODES

* Bot (#37)

* Add logtailLogger,
Fix bug in AuthServices.refresh

* Update CHANGELOG

* Update CONFIG.LOG.CODES

* Update USER_ALREADY_EXISTS error description

* Add on call bot flows,
Add endpoint to set on call status on/off

* Rename table BOT_MESSAGES to BOT_FLOWS

* Update REAMDE,
Update CHANGELOG

* Improve seed.ts

* Update CHANGELOG

* Add GET /agent/on-call endpoint

* Update README with on call endpoints

* Add tests for on-call endpoints

* Move test agent creds to .env

---------


---------

+++ src/helpers/notification.ts
@@ -1,20 +1,28 @@
+import { WebPushError } from "web-push";
 import { WebPushServices } from "@/components/web-push/services";
 import { WebPushPayload } from "@/types/request/web-push";
 import { WebPushDAO } from "@/db/web-push";
+import CONFIG from "@/config";
 
 export class Notify {
   static async agent(payload: WebPushPayload) {
-    const webPushServices = new WebPushServices();
-    const subscriptions = await webPushServices.index();
+    try {
+      const webPushServices = new WebPushServices();
+      const subscriptions = await webPushServices.index();
 
-    subscriptions.forEach(async (subscription) => {
-      try {
-        await webPushServices.send(subscription, payload);
-      } catch (e: any) {
-        if (e.statusCode === 404 || e.statusCode === 410) {
-          await WebPushDAO.deleteByEndpoint(e.endpoint);
+      subscriptions.forEach(async (subscription) => {
+        try {
+          await webPushServices.send(subscription, payload);
+        } catch (e: any) {
+          if (e instanceof WebPushError) {
+            if (e.statusCode === 404 || e.statusCode === 410) {
+              await WebPushDAO.deleteByEndpoint(e.endpoint);
+            }
+          }
         }
-      }
-    });
+      });
+    } catch (e) {
+      if (CONFIG.LOG.LEVEL === "debug") console.error(e);
+    }
   }
 }
